added validators and relations to entities throwing errors (Error creating bean with name 'transactionRepository', Not a managed type: class com.example.entity.Transaction)

+++ src/main/java/com/example/repository/AccountRepository.java
@@ -8,7 +8,8 @@ import org.springframework.stereotype.Repository;
 // Annotation
 @Repository
 public interface AccountRepository
-        extends JpaRepository<Account, Long> {
+        extends JpaRepository<Account, Integer> {
+
 
 
 }
